# CatalogoParametrosUcoLab/docs/openapi/catalogo-parametros-v1.yaml
openapi: 3.0.3
info:
  title: Catálogo de Parámetros UcoLab
  version: 1.0.0
  description: |
    Contrato público y versionado del módulo Catálogo de Parámetros.
    Este archivo es la fuente de verdad para productores y consumidores.
servers:
  - url: http://localhost:8080
    description: Desarrollo local
...
paths:
...
  /catalogo-parametros/api/v1/parametros:
    get:
      tags: [Parámetros]
      operationId: consultarParametros
      parameters: [{ $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/PageSize' }]
      responses:
        '200': { $ref: '#/components/responses/ParametrosOk' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      tags: [Parámetros]
      operationId: crearParametro
      requestBody: { $ref: '#/components/requestBodies/ParametroRequest' }
      responses:
        '201': { $ref: '#/components/responses/Created' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /catalogo-parametros/api/v1/parametros/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [Parámetros]
      operationId: consultarParametroPorId
      responses:
        '200': { $ref: '#/components/responses/ParametrosOk' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    put:
      tags: [Parámetros]
      operationId: actualizarParametro
      requestBody: { $ref: '#/components/requestBodies/ParametroRequest' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      tags: [Parámetros]
      operationId: eliminarParametro
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /catalogo-parametros/api/v1/parametros/{id}/changestatus:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      tags: [Parámetros]
      operationId: cambiarEstadoParametro
      summary: Cambia únicamente el estado activo del parámetro
      requestBody: { $ref: '#/components/requestBodies/CambiarEstadoRequest' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      description: Identificador UUID del recurso
      schema: { type: string, format: uuid }
    Page:
      name: page
      in: query
      description: Página basada en uno
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      description: Cantidad máxima de registros solicitados
      schema: { type: integer, minimum: 1, default: 10 }
  requestBodies:
    CambiarEstadoRequest:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/CambiarEstadoRequest' }
...
    ParametroRequest:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ParametroRequest' }
...
  responses:
    Success:
      description: Operación exitosa
      content: { application/json: { schema: { $ref: '#/components/schemas/MessageResponse' } } }
    Created:
      description: Recurso creado
      content: { application/json: { schema: { $ref: '#/components/schemas/MessageResponse' } } }
    BadRequest:
      description: Petición inválida o regla de negocio incumplida
      content: { application/json: { schema: { $ref: '#/components/schemas/MessageResponse' } } }
    NotFound:
      description: Recurso no encontrado
      content: { application/json: { schema: { $ref: '#/components/schemas/MessageResponse' } } }
    Conflict:
      description: Conflicto con el estado actual del catálogo
      content: { application/json: { schema: { $ref: '#/components/schemas/MessageResponse' } } }
    InternalError:
      description: Error técnico no esperado
      content: { application/json: { schema: { $ref: '#/components/schemas/MessageResponse' } } }
...
    ParametrosOk: { description: Parámetros encontrados, content: { application/json: { schema: { $ref: '#/components/schemas/ParametroResponse' } } } }
...
  schemas:
    CambiarEstadoRequest:
      type: object
      required: [activo]
      properties:
        activo: { type: boolean, example: false }
    Uuid:
      type: string
      format: uuid
      example: 123e4567-e89b-12d3-a456-426614174000
    InputUuid:
      type: string
      format: uuid
      description: Los DTO actuales reciben el UUID como cadena JSON.
      example: 123e4567-e89b-12d3-a456-426614174000
    InputBoolean:
      type: string
      enum: ['true', 'false']
      default: 'true'
      description: Los DTO actuales reciben el booleano como cadena JSON.
...
    MessageResponse:
      type: object
      required: [mensajes]
      properties:
        mensajes: { type: array, items: { type: string } }
...
    ParametroRequest:
      type: object
      additionalProperties: false
      required: [nombre, idFuncionalidad, idTipoParametro, activo]
      properties:
        nombre: { type: string, minLength: 1, example: idioma }
        idFuncionalidad: { $ref: '#/components/schemas/InputUuid' }
        idTipoParametro: { $ref: '#/components/schemas/InputUuid' }
        activo: { $ref: '#/components/schemas/InputBoolean' }
...
    Parametro:
      type: object
      required: [id, nombre, idFuncionalidad, idTipoParametro, activo]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        nombre: { type: string }
        idFuncionalidad: { $ref: '#/components/schemas/Uuid' }
        idTipoParametro: { $ref: '#/components/schemas/Uuid' }
        activo: { type: boolean }
...
    ParametroResponse: { allOf: [{ $ref: '#/components/schemas/MessageResponse' }, { type: object, required: [parametros], properties: { parametros: { type: array, items: { $ref: '#/components/schemas/Parametro' } } } }] }
...